Time Measure & Support for Multiple Measures (#285)
* Add simple time measurement

While imperfect, this is still useful data that anchors benchmark
iterations to a number that is more meaningful to humans sometimes.
Combined with the ability to capture multiple measures, this is
pretty useful to have available.

* Add support for capturing multiple measurements

Previously, only a single measure type could be requested.  Often,
we might want to capture multiple measures on a benchmark run as
different measurements might provide a more complete picture of
a given benchmark iteration.